Tried to make game object global, not working yet
I tried to make the game object global so that it doesn't have to be
passed as an argument to main.game_main(). That would make it easier to
create menus from YAML-files. However, it seems like there is some
problem with making the game-object global, probably because the
game_main() function is called from another module.