Music no longer restarts if the right song is playing
Until now, the music was reloaded every time a new menu or level was loaded, even if the currently playing song was supposed to play in the newly loaded menu/level. Now, a check is performed when loading music, and if the right song is already playing it doesn't reload it.