Tried to start using the camera class, not working yet
I tried to include the new camera object in the game, but there are still some problems, mainly with the rendering. The background isn't displayed on the right position, the objects aren't displayed on the right position and the screen isn't cleared between frames. All these things must be fixed asap.