Started working on adding fullscreen
It doesn't work as expected yet, a lot of bugs must be fixed. Some known
bugs added here:
* When in fullscreen in menus, the background is cropped so only a
portion of the wanted menu is shown on screen.
* The graphics in the levels don't work at all (I haven't started woring
on making them fullscreen compatible yet)
* Pop-up menus don't work, buttons etc are not visible. Unfortunately,
this means that it's impossible to quit the game without force-quitting
via the task manager. The reason that the menu items in pop-up menus
aren't shown is that their positions are calculated the wrong way
(further explanation as comment in source code).